Inline backend gRPC service into server binary
- Move `BackendEngineService` implementation from `service/backend.rs` into `bin/server.rs`
- Switch task creation to `task_id::mint_task_instance_id` and keep task key parsing centralized
- Update service module/proxy/tests to match the new backend service layout
