+++ engine/src/service/mod.rs
@@ -1,2 +1,1 @@
-pub mod backend;
 pub mod proxy;
